+++ archiveworker/quiz_archive_job.py
@@ -1,5 +1,5 @@
 # Moodle Quiz Archive Worker
-# Copyright (C) 2024 Niels Gandraß <niels@gandrass.de>
+# Copyright (C) 2025 Niels Gandraß <niels@gandrass.de>
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -218,7 +218,15 @@ class QuizArchiveJob:
         os.makedirs(f'{self.workdir}/attempts', exist_ok=True)
 
         async with async_playwright() as p:
-            browser = await p.chromium.launch(proxy={"server":"http://MYPROXYHOST:MYPROXYPORT"},args=['--disable-web-security'])  # Pass --disable-web-security to ignore CORS errors
+            browser = await p.chromium.launch(
+                args=['--disable-web-security'],  # Pass --disable-web-security to ignore CORS errors
+                proxy={
+                    'server': Config.PROXY_SERVER_URL,
+                    'username': Config.PROXY_USERNAME,
+                    'password': Config.PROXY_PASSWORD,
+                    'bypass': Config.PROXY_BYPASS_DOMAINS,
+                } if Config.PROXY_SERVER_URL else None
+            )
             context = await browser.new_context(viewport=ViewportSize(
                 width=int(Config.REPORT_BASE_VIEWPORT_WIDTH),
                 height=int(Config.REPORT_BASE_VIEWPORT_WIDTH / (16/9)))
@@ -583,7 +591,12 @@ class QuizArchiveJob:
             # Try to get JSON content if debug logging is enabled to allow debugging
             if Config.LOG_LEVEL == logging.DEBUG:
                 if content_type.startswith('application/json'):
-                    r = requests.get(url=download_url, params={'token': self.request.wstoken}, allow_redirects=True)
+                    r = requests.get(
+                        url=download_url,
+                        proxies=MoodleAPI.generate_proxy_settings(),
+                        params={'token': self.request.wstoken},
+                        allow_redirects=True
+                    )
                     self.logger.debug(f'Backup file GET response: {r.text}')
 
             # Normal error handling