+++ config.py
@@ -1,5 +1,5 @@
 # Moodle Quiz Archive Worker
-# Copyright (C) 2024 Niels Gandraß <niels@gandrass.de>
+# Copyright (C) 2025 Niels Gandraß <niels@gandrass.de>
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -40,6 +40,18 @@ class Config:
 
     SERVER_PORT = os.getenv('QUIZ_ARCHIVER_SERVER_PORT', default='8080')
     """Port for Flask to listen on"""
+
+    PROXY_SERVER_URL = os.getenv('QUIZ_ARCHIVER_PROXY_SERVER_URL', default=None)
+    """URL of the proxy server to use for all playwright requests. HTTP and SOCKS proxies are supported. If not set, auto-detection will be performed. If set to false, no proxy will be used."""
+
+    PROXY_USERNAME = None
+    """Optional username to authenticate at the proxy server. Will be populated based on PROXY_SERVER_URL."""
+
+    PROXY_PASSWORD = None
+    """Optional password to authenticate at the proxy server. Will be populated based on PROXY_SERVER_URL."""
+
+    PROXY_BYPASS_DOMAINS = os.getenv('QUIZ_ARCHIVER_PROXY_BYPASS_DOMAINS', default=None)
+    """Comma-separated list of domains that should always be accessed directly, bypassing the proxy"""
 
     QUEUE_SIZE = int(os.getenv('QUIZ_ARCHIVER_QUEUE_SIZE', default=8))
     """Maximum number of requests that are queued before returning an error."""